First attempt to break too wide tables into multiple ones
WARNING: This has been done REALLY ugly and only for my specific
use case. School book example of a quick & dirty implementation.

Introduced new "headerColumns" setting for tables so you can configure
how many of the most left columns should be copied with to the next table.
Just as the "headerRows" option, but for columns (obiously).